# [gputests] Don't use temporary datapath in test_torchscript